+++ Tests/StreamChatUITests/SnapshotTests/ChatMessageList/ChatMessage/ChatMessageContentView_Tests.swift
@@ -277,6 +277,41 @@ final class ChatMessageContentView_Tests: XCTestCase {
 
         AssertSnapshot(view, variants: .onlyUserInterfaceStyles)
     }
+    
+    func test_appearance_whenMessageWithMarkdownItalicStyleFromTheCurrentUserIsSent() {
+        let channelWithReadsEnabled: ChatChannel = .mock(
+            cid: .unique,
+            config: .mock(readEventsEnabled: true)
+        )
+
+        let messageWithMarkdown = """
+           https://getstream.io/test_endpoint_in__snake__style
+           Test_text_in__snake__style and plain text
+           Test plain text and the_text_in__snake__style
+           _Test text in italic style_ and plain text
+           Test plain text and _text in italic style_
+           Test plain text, _text in italic style_ and plain text
+           Test plain text, __text in bold style__ and plain text
+        """
+
+        let sentMessageFromCurrentUser: ChatMessage = .mock(
+            id: .unique,
+            cid: channelWithReadsEnabled.cid,
+            text: messageWithMarkdown,
+            author: me,
+            createdAt: createdAt,
+            localState: nil,
+            isSentByCurrentUser: true,
+            readBy: []
+        )
+
+        let view = contentView(
+            message: sentMessageFromCurrentUser,
+            channel: channelWithReadsEnabled
+        )
+
+        AssertSnapshot(view, variants: .onlyUserInterfaceStyles)
+    }
 
     func test_appearance_whenMessageFromTheCurrentUserIsRead_inDirectMesssagesChannel() {
         let dmChannelWithReadsEnabled: ChatChannel = .mock(